+++ cmd/anubis/policy.go
@@ -157,9 +157,11 @@ func cr(name string, rule config.Rule) CheckResult {
 	}
 }
 
+// checkRemoteAddress checks if the given IP address is within the expected range of the bot,
+// or if none applies.
 func (s *Server) checkRemoteAddress(b Bot, addr net.IP) bool {
 	if b.Ranger == nil {
-		return false
+		return true
 	}
 
 	ok, err := b.Ranger.Contains(addr)
@@ -184,13 +186,13 @@ func (s *Server) check(r *http.Request) (CheckResult, *Bot, error) {
 
 	for _, b := range s.policy.Bots {
 		if b.UserAgent != nil {
-			if uaMatch := b.UserAgent.MatchString(r.UserAgent()); uaMatch || (uaMatch && s.checkRemoteAddress(b, addr)) {
+			if b.UserAgent.MatchString(r.UserAgent()) && s.checkRemoteAddress(b, addr) {
 				return cr("bot/"+b.Name, b.Action), &b, nil
 			}
 		}
 
 		if b.Path != nil {
-			if pathMatch := b.Path.MatchString(r.URL.Path); pathMatch || (pathMatch && s.checkRemoteAddress(b, addr)) {
+			if b.Path.MatchString(r.URL.Path) && s.checkRemoteAddress(b, addr) {
 				return cr("bot/"+b.Name, b.Action), &b, nil
 			}
 		}